+++ src/frequenz/channels/_broadcast.py
@@ -15,6 +15,7 @@ from uuid import UUID, uuid4
 from ._base_classes import ChannelClosedError
 from ._base_classes import Peekable as BasePeekable
 from ._base_classes import Receiver as BaseReceiver
+from ._base_classes import ReceiverError, ReceiverStoppedError
 from ._base_classes import Sender as BaseSender
 from ._base_classes import SenderError, T
 
@@ -256,11 +257,11 @@ class Receiver(BaseReceiver[T]):
         """Wait until the receiver is ready with a value.
 
         Raises:
-            EOFError: if this receiver is no longer active.
-            ChannelClosedError: if the underlying channel is closed.
+            ReceiverStoppedError: if there is some problem with the receiver.
+            ReceiverError: if the receiver is not longer active.
         """
         if not self._active:
-            raise EOFError("This receiver is no longer active.")
+            raise ReceiverError("This receiver is no longer active.", self)
 
         # Use a while loop here, to handle spurious wakeups of condition variables.
         #
@@ -268,7 +269,7 @@ class Receiver(BaseReceiver[T]):
         # consumed, then we return immediately.
         while len(self._q) == 0:
             if self._chan.closed:
-                raise ChannelClosedError()
+                raise ReceiverStoppedError(self) from ChannelClosedError(self._chan)
             async with self._chan.recv_cv:
                 await self._chan.recv_cv.wait()